requirements.txt
streamlit
ultralytics
Pillow
tempfile
opencv-python
numpy
pytesseract
